added bit shift operations

+++ app/src/test/java/com/annushkaproject/programmerscalculator/ProgrammerSignChangeTest.java
@@ -27,20 +27,20 @@ public class ProgrammerSignChangeTest {
     }
     @Test
     public void programmerSignChangeWORD_isCorrect() {
-        assertEquals(-5, ProgrammerUtil.changeSign(WORD, 5));
-        assertEquals(5, ProgrammerUtil.changeSign(WORD, -5));
-        assertEquals((short) 0xC000, ProgrammerUtil.changeSign(WORD, (short) 0x4000));
-        assertEquals((short) 0x8000, ProgrammerUtil.changeSign(WORD, (short) 0x8000));
-        assertEquals((short) 0x8001, ProgrammerUtil.changeSign(WORD, (short) 0x7FFF));
-        assertEquals(1, ProgrammerUtil.changeSign(WORD, (short) 0xFFFF));
+        assertEquals((short) -5, ProgrammerUtil.changeSign(WORD, 5));
+        assertEquals((short) 5, ProgrammerUtil.changeSign(WORD, -5));
+        assertEquals((short) 0xC000, ProgrammerUtil.changeSign(WORD,0x4000));
+        assertEquals((short) 0x8000, ProgrammerUtil.changeSign(WORD,0x8000));
+        assertEquals((short) 0x8001, ProgrammerUtil.changeSign(WORD,0x7FFF));
+        assertEquals((short) 1, ProgrammerUtil.changeSign(WORD,0xFFFF));
     }
     @Test
     public void programmerSignChangeBYTE_isCorrect() {
-        assertEquals(-5, ProgrammerUtil.changeSign(BYTE, 5));
-        assertEquals(5, ProgrammerUtil.changeSign(BYTE, -5));
-        assertEquals((byte) 0xC0, ProgrammerUtil.changeSign(BYTE, (byte) 0x40));
-        assertEquals((byte) 0x80, ProgrammerUtil.changeSign(BYTE, (byte) 0x80));
-        assertEquals((byte) 0x81, ProgrammerUtil.changeSign(BYTE, (byte) 0x7F));
-        assertEquals(1, ProgrammerUtil.changeSign(BYTE, (byte) 0xFF));
+        assertEquals((byte) -5, ProgrammerUtil.changeSign(BYTE, 5));
+        assertEquals((byte) 5, ProgrammerUtil.changeSign(BYTE, -5));
+        assertEquals((byte) 0xC0, ProgrammerUtil.changeSign(BYTE,0x40));
+        assertEquals((byte) 0x80, ProgrammerUtil.changeSign(BYTE,0x80));
+        assertEquals((byte) 0x81, ProgrammerUtil.changeSign(BYTE,0x7F));
+        assertEquals((byte) 1, ProgrammerUtil.changeSign(BYTE,0xFF));
     }
 }
